smaller cleanups
