+++ commands/db/host/start-task.py
@@ -1,45 +1,14 @@
 import os
 
-import discord
 from discord.ext import commands
 from dotenv import load_dotenv
 
-from api.errors import NoSubmissionChannelError
-from api.submissions import get_submission_channel
 from api.task_handling import start_task
-from api.utils import has_host_role, get_submitter_role
+from api.utils import has_host_role
+from discord_ext.discord_utils import delete_previous_current_submissions, remove_submitter_role
 
 load_dotenv()
 DEFAULT = os.getenv('DEFAULT')
-
-async def remove_submitter_role(bot, guild_id: int) -> None:
-    # Also clear submitter roles
-    submitter_role = await get_submitter_role(DEFAULT)
-    server = bot.get_guild(guild_id)
-    role = server.get_role(submitter_role)
-
-    for member in role.members:
-        try:
-            await member.remove_roles(role)
-        except discord.Forbidden:
-            print(
-                f"Failed to remove role {role.name} from {member.display_name} due to insufficient permissions.")
-        except discord.HTTPException as e:
-            print(f"Failed to remove role {role.name} from {member.display_name} due to an error: {e}")
-
-async def delete_previous_current_submissions(bot) -> None:
-    # Delete previous "Current submissions" message in submission channel
-    channel_id = await get_submission_channel(DEFAULT)
-    channel = bot.get_channel(channel_id)
-
-    try:
-        async for message in channel.history(limit=5):  # Adjust limit as necessary
-            if message.author == bot.user:
-                await message.delete()
-                break
-
-    except AttributeError:
-        raise NoSubmissionChannelError("Please set the submission channel with `/set-submission-channel`! (Ask an admin if you do not have permission)")
 
 
 class Start(commands.Cog):
@@ -51,20 +20,19 @@ class Start(commands.Cog):
     async def command(self, ctx, number: int, team_size: int = 1, multiple_tracks: int = 0,
                       speed_task: int = 0, year: int = None, deadline: int = None) -> None:
         await ctx.defer()
-        result = await start_task(number, team_size, multiple_tracks,
+        await start_task(number, team_size, multiple_tracks,
                                   speed_task, year, deadline, ctx.guild.id, ctx.message.guild.id)
-        if result:
-            # Delete previous "Current submissions" message in submission channel
-            await delete_previous_current_submissions(self.bot)
+        # Delete previous "Current submissions" message in submission channel
+        await delete_previous_current_submissions(self.bot)
 
-            # Successful message to send:
-            if deadline is not None:
-                await ctx.send(f"Successfully started **Task {number} - {year}**! Deadline: <t:{deadline}:f>")
-            else:
-                await ctx.send(f"Successfully started **Task {number} - {year}**!")
+        # Successful message to send:
+        if deadline is not None:
+            await ctx.send(f"Successfully started **Task {number} - {year}**! Deadline: <t:{deadline}:f>")
+        else:
+            await ctx.send(f"Successfully started **Task {number} - {year}**!")
 
-            # Remove the submitter role from all members
-            await remove_submitter_role(self.bot, ctx.guild.id)
+        # Remove the submitter role from all members
+        await remove_submitter_role(self.bot, ctx.guild.id)
 
         ######################
         #### TEAMS SYSTEM ####
@@ -104,6 +72,7 @@ class Start(commands.Cog):
         #
         # except AttributeError:
         #     await ctx.send(f"Please set the seeking channel with `/set-seeking-channel`! (Ask an admin if you do not have permission)")
+
 
 async def setup(bot) -> None:
     await bot.add_cog(Start(bot))
